refactor(exceptions): reorganize exceptions into domain packages
Move exceptions from shared package to their respective domain packages
(account, transaction, security, user) to improve code organization and
follow domain-driven design principles. Update import statements across
the codebase to reflect new exception locations.
