+++ src/main/java/info/mackiewicz/bankapp/account/exception/AccountNotFoundByIbanException.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.shared.exception;
+package info.mackiewicz.bankapp.account.exception;
 
 public class AccountNotFoundByIbanException extends RuntimeException {
     private static final String DEFAULT_MESSAGE = "Account not found for the provided IBAN";
