+++ src/main/java/info/mackiewicz/bankapp/account/exception/AccountNotFoundByIdException.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.shared.exception;
+package info.mackiewicz.bankapp.account.exception;
 
 public class AccountNotFoundByIdException extends RuntimeException {
     private static final String DEFAULT_MESSAGE = "Account not found for the provided ID";
