+++ src/main/java/info/mackiewicz/bankapp/presentation/dashboard/service/SettingsService.java
@@ -3,8 +3,8 @@ package info.mackiewicz.bankapp.presentation.dashboard.service;
 import info.mackiewicz.bankapp.presentation.dashboard.dto.ChangePasswordRequest;
 import info.mackiewicz.bankapp.presentation.dashboard.dto.ChangeUsernameRequest;
 import info.mackiewicz.bankapp.presentation.dashboard.dto.UserSettingsDTO;
+import info.mackiewicz.bankapp.presentation.exception.InvalidUserException;
 import info.mackiewicz.bankapp.security.service.PasswordService;
-import info.mackiewicz.bankapp.shared.exception.InvalidUserException;
 import info.mackiewicz.bankapp.user.model.User;
 import info.mackiewicz.bankapp.user.service.UserService;
 import lombok.RequiredArgsConstructor;
