+++ src/main/java/info/mackiewicz/bankapp/security/service/PasswordResetTokenService.java
@@ -6,9 +6,9 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import info.mackiewicz.bankapp.security.exception.TooManyPasswordResetAttemptsException;
 import info.mackiewicz.bankapp.security.model.PasswordResetToken;
 import info.mackiewicz.bankapp.security.repository.PasswordResetTokenRepository;
-import info.mackiewicz.bankapp.shared.exception.TooManyPasswordResetAttemptsException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
