+++ src/main/java/info/mackiewicz/bankapp/transaction/exception/TransactionAlreadyProcessedException.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.shared.exception;
+package info.mackiewicz.bankapp.transaction.exception;
 
 public class TransactionAlreadyProcessedException extends RuntimeException {
     private static final String DEFAULT_MESSAGE = "Transaction has already been processed";
