+++ src/main/java/info/mackiewicz/bankapp/transaction/exception/TransactionCannotBeProcessedException.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.shared.exception;
+package info.mackiewicz.bankapp.transaction.exception;
 
 public class TransactionCannotBeProcessedException extends RuntimeException {
     private static final String DEFAULT_MESSAGE = "Transaction cannot be processed";
