+++ src/main/java/info/mackiewicz/bankapp/transaction/exception/TransactionNotFoundException.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.shared.exception;
+package info.mackiewicz.bankapp.transaction.exception;
 
 public class TransactionNotFoundException extends RuntimeException {
     private static final String DEFAULT_MESSAGE = "Transaction not found";
