+++ src/main/java/info/mackiewicz/bankapp/transaction/exception/TransactionSourceAccountNotSpecifiedException.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.shared.exception;
+package info.mackiewicz.bankapp.transaction.exception;
 
 public class TransactionSourceAccountNotSpecifiedException extends RuntimeException {
     private static final String DEFAULT_MESSAGE = "Transaction source account must be specified";
