+++ src/main/java/info/mackiewicz/bankapp/transaction/model/builder/DepositBuilder.java
@@ -1,7 +1,7 @@
 package info.mackiewicz.bankapp.transaction.model.builder;
 
 import info.mackiewicz.bankapp.account.model.Account;
-import info.mackiewicz.bankapp.shared.exception.TransactionDestinationAccountNotSpecifiedException;
+import info.mackiewicz.bankapp.transaction.exception.TransactionDestinationAccountNotSpecifiedException;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
 import info.mackiewicz.bankapp.transaction.model.TransactionType;
