+++ src/main/java/info/mackiewicz/bankapp/transaction/model/builder/WithdrawalBuilder.java
@@ -1,7 +1,7 @@
 package info.mackiewicz.bankapp.transaction.model.builder;
 
 import info.mackiewicz.bankapp.account.model.Account;
-import info.mackiewicz.bankapp.shared.exception.TransactionSourceAccountNotSpecifiedException;
+import info.mackiewicz.bankapp.transaction.exception.TransactionSourceAccountNotSpecifiedException;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
 import info.mackiewicz.bankapp.transaction.model.TransactionType;
