+++ src/main/java/info/mackiewicz/bankapp/user/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.shared.exception;
+package info.mackiewicz.bankapp.user.exception;
 
 public class UserNotFoundException extends RuntimeException {
     private static final String DEFAULT_MESSAGE = "User not found";
